+++ src/pages/Legal.tsx
