+++ src/pages/Terms.tsx
